Compatibility with ConfigParser on python 3.14
In 3.14, the ConfigParser was pretty substantially reworked. This patch
modifies the SystemdUnitParser to better handle calling into the
RawConfigParser base class (allowing us to pass arguments up to it) and
also provides for a set of fallback comment delimiters.

Note that it doees not attempt to reimplement the _read() function to be
in line with the modern ConfigParser. That will require a significant
rewrite at a later date.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>